clean up, leak squashed function count adjusted, folder structure adjusted.
