+++ sources/exec/exec.c
@@ -6,44 +6,11 @@
 /*   By: minsunki <minsunki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/03 22:24:00 by minsunki          #+#    #+#             */
-/*   Updated: 2022/03/07 00:27:48 by minsunki         ###   ########seoul.kr  */
+/*   Updated: 2022/03/07 01:26:01 by minsunki         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-// TODO 대소문자 구분 해야하나?
-
-static t_token	*next_cmd(t_token *tok)
-{
-	while (tok)
-	{
-		if (token_ident(tok) == T_CMD && !(tok->type & (REDIR)))
-			return (tok);
-		tok = tok->next;
-	}
-	return (0);
-}
-
-static int	is_builtin(char *bin)
-{
-	if (ft_strcmp(bin, "cd") == 0)
-		return (B_CD);
-	else if (ft_strcmp(bin, "echo") == 0)
-		return (B_ECHO);
-	else if (ft_strcmp(bin, "env") == 0)
-		return (B_ENV);
-	else if (ft_strcmp(bin, "exit") == 0)
-		return (B_EXIT);
-	else if (ft_strcmp(bin, "export") == 0)
-		return (B_EXPT);
-	else if (ft_strcmp(bin, "pwd") == 0)
-		return (B_PWD);
-	else if (ft_strcmp(bin, "unset") == 0)
-		return (B_UNST);
-	else
-		return (0);
-}
 
 static int	execute_builtin(t_meta *m, int flag)
 {
@@ -78,9 +45,6 @@ static int	execute_bin(t_meta *m)
 		ret = bin_run(m, m->argv[0]);
 	else
 		ret = bin_run(m, bin);
-	// if (!bin)
-	// 	return (1); // TODO MANAGE ERROR CODES. FILE NOT FOUND
-	// ret = bin_run(m, bin);
 }
 
 static void	exec_cmd(t_meta *m, t_token *tok)
@@ -90,10 +54,6 @@ static void	exec_cmd(t_meta *m, t_token *tok)
 	if (m->stop)
 		return ;
 	m->stop = 1;
-	// if (m->child)
-	// 	ft_putstr_fd("child ",STDERR_FILENO);
-	// ft_putstr_fd("exec cmd: ", STDERR_FILENO);
-	// ft_putendl_fd(tok->str, STDERR_FILENO);
 	m->argv = argv_build(tok);
 	bi = is_builtin(m->argv[0]);
 	if (bi)
@@ -102,10 +62,6 @@ static void	exec_cmd(t_meta *m, t_token *tok)
 		execute_bin(m);
 	argv_destroy(m);
 }
-// echo > a
-// a 실행 후 echo 실행
-// pipe 를 기준으로 후위순회 
-// echo test | cat
 
 static void	execute(t_meta *m, t_token *tok)
 {
@@ -116,9 +72,6 @@ static void	execute(t_meta *m, t_token *tok)
 	pt = token_prev_delim(tok);
 	nt = token_next_delim(tok);
 	piped = 0;
-	// ft_putstr_fd("EXECUTE ",STDERR_FILENO);
-	// ft_putendl_fd(tok->str, STDERR_FILENO);
-
 	if (pt && (pt->type & (T_RDR | T_APR)))
 		redir_r(m, tok, pt->type);
 	else if (pt && (pt->type & (T_RDL | T_APL)))
@@ -140,7 +93,6 @@ void	exec_start(t_meta *m)
 	execute(m, ct);
 	dup2(m->stdin,STDIN_FILENO);
 	dup2(m->stdout,STDOUT_FILENO);
-	// printf("%d %d\n",m->fd_out, m->fd_in);
 	fd_close(m->fd_out);
 	fd_close(m->fd_in);
 	waitpid(-1, &stat, 0);
@@ -148,36 +100,4 @@ void	exec_start(t_meta *m)
 	if (m->child)
 		mexit(stat);
 	m->stop = 0;
-	// ct = next_cmd(m->token_start);
-	// execute(m, ct);
-	// dup2(m->stdin,STDIN_FILENO);
-	// dup2(m->stdout,STDOUT_FILENO);
-	// close(m->fd_out);
-	// close(m->fd_in);
-	// waitpid(-1, &stat, 0);
-	// stat = WEXITSTATUS(stat);
-	// if (m->child)
-	// {
-	// 	ft_putendl_fd("EXIT HERE",STDERR_FILENO);
-	// 	// printf("exit here\n");
-	// 	mexit(stat);
-	// }
-	// m->stop = 0;
-	// ft_putendl_fd("WATSUP!",STDERR_FILENO);
-	// ct = next_cmd(m->token_start->next);
-	// ct->prev = 0;
-	// execute(m, ct);
-	// waitpid(-1, &stat, 0);
-	// stat = WEXITSTATUS(stat);
-	// if (m->child)
-	// {
-	// 	printf("exit here\n");
-	// 	mexit(stat);
-	// }
-	// while (ct)
-	// {
-	
-		// ct = next_cmd(m->token_start->next);
-		// break ;
-	// }
 }
